requirements.txt
pandas==2.2.3
PyYAML==6.0.2
elasticsearch==8.15.1
sqlite-utils==3.36
jupyter==1.0.0
